Extract drag state to new file
